ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
requests==2.31.0
numpy==1.26.2
python-multipart==0.0.6
aiohttp==3.9.1
imagehash==4.3.1
